refactor: dual-format shared package, deduplicate writeRuntimeSessionHint, consolidate buildMergedMcpConfig
- Shared package now emits both ESM (dist/) and CJS (dist/cjs/) via
  dual tsconfig builds. Fixes TS1479 warnings when the CJS API package
  imports from shared — TypeScript resolves the require condition to the
  CJS output and its co-located .d.ts files. The ESM root stays for CLI
  and other ESM consumers. This is necessary because the API uses
  __dirname extensively and can't easily migrate to ESM.
- writeRuntimeSessionHint: moved from claude-runner.ts and codex-runner.ts
  (identical 55-line copies) into packages/shared/src/runner/runtime-hints.ts.
  Both runners now import from shared.
- buildMergedMcpConfig: refactored to delegate session header injection to
  shared injectSessionHeaders(), then layers CLI-only skill server merging
  on top. Same logic, no duplication with server-side header injection.
- Removed unused fs/path imports from claude-runner after deduplication.
